requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
urllib3==2.2.1

# Text processing
nltk==3.8.1
spacy==3.7.4
regex==2023.12.25

# Embeddings
sentence-transformers==2.5.1
transformers==4.38.2
torch==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Vector storage
faiss-cpu

# LLM integration
openai==1.12.0

# Data handling
pandas

# Web API (optional)
# fastapi==0.109.0
# uvicorn==0.25.0
# pydantic==2.5.3

# Web UI (optional)
# flask==3.0.0
# streamlit==1.31.0  # Optional for quick UI

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
pytest==7.4.3  # For testing
jupyter==1.0.0  # For notebooks
matplotlib==3.8.2  # For visualizations

ollama-python
anthropic
